refactor: renamed to authService
